refactor: ドメインロジックを models/ ディレクトリに集約
- components/session/streamParsing.ts → models/stream.ts
- components/session/toolHelpers.ts → models/tool.ts
- SessionDetail.tsx 内の TodoItem 型と parseTodoInput を models/tool.ts に移動
