+++ frontend/src/models/stream.ts
